tests/wpt/tests/html/interaction/focus/focusgroup/tentative/sequential-navigation/directional-input-handler-multiple-in-segment.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focusgroup - Multiple native directional input handlers in same segment</title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com/">
<link rel="help" href="https://open-ui.org/components/scoped-focusgroup.explainer/">
<meta name="assert" content="Multiple native directional input handlers can exist in a segment, each blocks directional navigation independently.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/shadow-dom/focus-navigation/resources/focus-utils.js"></script>
<script src="../resources/focusgroup-utils.js"></script>

<div id="toolbar" focusgroup="toolbar">
  <button tabindex=0 id="btn1">Button 1</button>
  <input id="input1" type="text" placeholder="First">
  <button tabindex=0 id="btn2">Button 2</button>
  <select id="select1">
    <option>Option 1</option>
    <option>Option 2</option>
  </select>
  <button tabindex=0 id="btn3">Button 3</button>
</div>

<script>

  promise_test(async t => {
    const btn1 = document.getElementById("btn1");
    const input1 = document.getElementById("input1");
    const btn2 = document.getElementById("btn2");
    const select1 = document.getElementById("select1");

    // Directional navigation TO native directional input handlers works.
    await focusAndSendDirectionalInput(btn1, kRight);
    assert_equals(document.activeElement, input1);
    await focusAndSendDirectionalInput(btn2, kRight);
    assert_equals(document.activeElement, select1);

    // Each native directional input handler blocks directional navigation when focused.
    await assert_directional_input_does_not_move_focus(input1);
    await assert_directional_input_does_not_move_focus(select1);
  }, "Multiple native directional input handlers each block directional navigation independently");

</script>
